+++ src/main/java/com/quant/craft/backend/application/service/AuthService.java
@@ -7,8 +7,8 @@ import com.quant.craft.backend.infrastructure.client.auth.OAuthClient;
 import com.quant.craft.backend.infrastructure.client.auth.OAuthClientFactory;
 import com.quant.craft.backend.infrastructure.client.auth.dto.UserResponse;
 import com.quant.craft.backend.infrastructure.repository.UserRepository;
-import com.quant.craft.backend.presentation.controller.dto.auth.AuthorizationCodeRequest;
-import com.quant.craft.backend.presentation.controller.dto.auth.TokenResponse;
+import com.quant.craft.backend.presentation.controller.auth.dto.AuthorizationCodeRequest;
+import com.quant.craft.backend.presentation.controller.auth.dto.TokenResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
