+++ src/main/java/com/quant/craft/backend/application/service/UserService.java
@@ -3,7 +3,7 @@ package com.quant.craft.backend.application.service;
 import com.quant.craft.backend.domain.user.User;
 import com.quant.craft.backend.exception.NotFoundException;
 import com.quant.craft.backend.infrastructure.repository.UserRepository;
-import com.quant.craft.backend.presentation.controller.dto.user.UserResponse;
+import com.quant.craft.backend.presentation.controller.user.dto.UserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
